Move TargetWidgetsInvertingSelectionListener to util.selection
